Fix #10091, which-key interface delay/ lag.
A fix for #10091. The default behavour for which-key is for
'which-key-idle-delay' to affect both the initial trigger and subsequent
actions. By setting 'which-key-idle-secondary-delay' to something non-nil,
that delay acts over the subsequent actions instead.

The which-key wiki recommends setting it to a 'non-zero value', as zero could
cause issues, so instead set it to 0.01.