Reset useful and useless buffers to empty lists
I feel like we should start from scratch on this one and carefully choose the
defaults.

Also this settings is a very personal setting so if we make some buffers
useless we must have a consensus on it. Marking all special buffers starting
with `*` as useless is too aggressive and make Spacemacs less POLA since two
consecutive press on SPC TAB may not revert to the original buffer.