jabber: Fix missing optional argument for the alert jabber-alert-echo function.
Jabber alert function should accept an optional second argument containing a
title of the message. The function mimic the default implementation of
jabber-alert-echo by showing the title if it is present and a message body
otherwise.

This commit fixes an error which is thrown when jabber receives a new notification:
(wrong-number-of-arguments (lambda (msg) (if (minibuffer-prompt) nil (message "%s" msg))) 2)