handle .venv created by pipenv correctly
when pipenv is run as...

PIPENV_VENV_IN_PROJECT=1 pipenv ...

a virtualenv directory called `.venv` is installed in the directory as the
`Pipfile`

the current functionality is to look for a file and then tell pyenv to 'workon'
it's contents. doing the above will cause it to blow up.

this change checks to see if it's a directory or file and either calls

pyvenv-activate if directory

or

pyvenv-workon with the contents of the file if file(existing behaviour)

then end result, a virtualenv created local for the project by pipenv will get
activated automatically instead of blowing up like it does now