Add paths to exec-path in correct order
`add-to-list` prepends items to a list by default.  So, calling it sequentially
on directories in $PATH results in the user's directories appearing in
`exec-path` in the opposite order in which the appear in her $PATH.  This can
result in unexpected behavior.

Processing the directories in $PATH in reverse order resolves this issue.