Clojure layer: fix function usage in clojure layer according to cider changes
The function `cider-current-repl-buffer` does not exist anymore in cider after PR [#2324](https://github.com/clojure-emacs/cider/pull/2324) was merged.
It broke the usage of functions in the clojure layer that sent used `spacemacs//cider-eval-in-repl-no-focus`, which sends forms to be evaluated in the repl buffer without focusing on it.
The fix is to use `cider-current-repl` instead of `cider-current-repl-buffer`