first pass at generic-eval-region and generic-send-region
First, this is a Work In Progress PR intended to gauge whether this
feature is still wanted and how I should go about it. This work started
life based on #3103 and my gist in that thread.

This provides two primary functions which abstract the idea of evaling
and sending a region to a repl or process. This is done to deduplicate
mode-specific eval and send functions. The idea is that instead of
building a special your-mode-eval-region, your-mode-eval-buffer,
your-mode-eval-until-end-of-line, your-model-eval-defun, we can simply
implement a function like:

(defun your-mode-eval-region (beg end)
   "Evals for your-mode"
   (interactive)
   .....)
And then hook it onto spacemacs--generic-eval-region-alist like so:

(add-to-list 'spacemacs--generic-eval-region-alist '(your-mode .
your-mode-eval-region))
In exchange, you get:

An evil operator for evaling.
Functions for evaling buffers, defuns, paragraphs, active marked
regions, current line, etc.
Optional highlighting and inline-results via eval-sexp-fu.
Consistency behavior and capabilities across modes.
This code makes a distinction between evaling and sending. Eval is when
you can talk to a repl or process and get results back. Send is when
you're limited to sending text to some other buffer which is hopefully
running a REPL.

This WIP is intended to discover:

if this is wanted (I've been using my own generic-eval.el for several
years now, so I know I like it.)
how to break apart this into reviewable chunks
what documentation needs to be created
how badly written this code is (emacs lisp isn't something I use a ton)
So far, this WIP does not cover:

mode/layer-specific modifications to use these new commands
My intention with not including changes to layers is to be able to shake
out the usefulness of this to the community at large before beginning
negotiations to rip out a ton of custom, mode-specific eval/send
functions.