Fix font loading issue on first instance of emacsclient
If you start an emacsclient when no frame is currently available, font loading
will not run without a call to message. Font setting will become correct when
the second emacsclient is started. This was discovered and fixed in commit
 #04267777, and was broken in commit #ac247396, where a call to message is
replaced with a call to spacemacs-buffer/message, which will only print this
message in debug mode.