Fix compilation finish hook
Use compilation-finish-functions instead of compilation-finish-function.

Spacemacs assigns an anonymous function to the compilation-finish-function
variable, which was deprecated in Emacs 22.1 and removed in Emacs 27.1 in
favor of the compilation-finish-functions hook.  This commit deletes this
assignment, replaces the anonymous function with a named function, and adds
the newly named function to compilation-finish-functions.

* layers/+spacemacs/spacemacs-defaults/funcs.el
(compilation-finish-function): Delete assignment of an anonymous function.
(spacemacs/compilation-finish-function): New defun based on the old
anonymous function.
* layers/+spacemacs/spacemacs-defaults/config.el
(compilation-finish-functions): Add spacemacs/compilation-finish-function.