Introduces Dart layer
This is a initial implementation of a Dart later.
A [previous implementation] has been deprecated, using a forked version of `dart-mode`.

This layer uses the [dart-mode] linked on the [Dart tools] page.
The layer works with a Flutter installation as well.

Ideally, we would be able to offer a `lsp` backend as an option to use, instead
of using the Dart Analyzer, but I'm still figuring out how to implement and
integrate this.

This layer has [some] [fixes] to make it work, and it is waiting for upstream to
take patches.

- Resolves https://github.com/syl20bnr/spacemacs/issues/8333
- Resolves https://github.com/syl20bnr/spacemacs/issues/8395
- Resolves https://github.com/syl20bnr/spacemacs/issues/10452

---
[previous implementation]: https://github.com/syl20bnr/spacemacs/pull/8406
[dart-mode]: https://github.com/bradyt/dart-mode
[Dart tools]: https://www.dartlang.org/tools
[some]: https://github.com/bradyt/dart-mode/pull/67
[fixes]: https://github.com/bradyt/dart-mode/pull/66