Move lexical-binding to first line
problem:
Some files write the lexical binding to the right of the first line, that makes
it seem like it's just part of the layer description.

https://www.emacswiki.org/emacs/LexicalBinding states that:
> enable lexical binding in your elisp library, put a line like this at the top
> of the file:
> ;; -*- lexical-binding: t -*-

which seems to indicate that it should be written by itself.

And possibly the newest layer multiple-cursors writes the lexical binding on the
first line.

solution:
Move the lexical-binding to the first line.