evilified-state: Restore normal state keymap on exit
When a buffer in evil-evilified-state switches to evil-normal-state, the
keymap in the evil-normal-state has then a peculiar mapping of an
escape key that, when pressed, switches the buffer back to the
evil-evilified-state.

This behavior is not very convenient, for example, when you use the
debug mode for debugging your code. During the debug session, the
buffers with debugged source code are in evil-evilified-state and after
finishing the debugging all those buffers suffer from the escape
returning to evil-evilified-state issue described above.

Restoring the saved normal state keymap when exiting the evilified state
solves the issue.