Increase compatibility with guix
* Variable elpa-root-directory
spacemacs tries to install elpa packages to spacemacs-start-directory.
On guix spacemacs-start-directory is a store location, so we set it to
user-emacs-dir. This is the location that it gets installed to on
vanilla emacs.

* Template generation
When spacemacs-install runs on first boot it copies a template file
from spacemacs-start-directory which causes it to be read-only. This
causes an error when spacemacs tries to edit the template based on user
input. We inhibit read-only after opening the template file. This works
because the the file is written into dotspacemacs-directory, not to the
location it was opened from.

* Variable spacemacs-start-directory
spacemacs assumes that spacemacs is installed into `$HOME/.emacs.d`. We
need to update it to point to the location in the guix store. Since the
variable is defined in spacemacs/core/core-load-paths.el we set it to
the second directory above `load-file-name`

* Quelpa
Some quelpa functions do not work on work on spacemacs-for-guix
out of the box. When they copy files from the store the files are
left as read-only. The only change is to set file permissions after
copying a file or directory into Quelpa's build directory,
~/.emacs.d/.cache/quelpa/build