[clojure] Fix jumping backwards after cider-find-var
`cider-jump-to-var` was renamed to `cider-find-var` here:
https://github.com/clojure-emacs/cider/commit/ceb2d7df75f09c35ef17a7ac050cd8c4388d32ce#diff-e39c33031dc8d3e78b32ebee83293cd6L809-R846

In https://github.com/syl20bnr/spacemacs/pull/9792/ `spacemacs/clj-find-var`
was added to be used as the jump handler for clojure modes. It uses
`cider-find-var` which doesn't push jump positions onto the evil jump stack so
the no longer current defadvice was more apparent.