Fix home buffer version and banner
- Removed the first line deletion.
This required the banners to have an empty first line.

- Removed the empty first line from the banners.

- Made sure that there is a single empty line between:
  - The Version and Banner.
  - Banner and Buttons ([?] [Homepage] ...).
  - Version and Buttons (when the banner is hidden).

- Reordered the backend version insertion function call
to match the frontend Spacemacs buffer order.
  - version
  - banner
  - buttons

Before it added:
- the banner
- then the buffers first line was removed
- then the version was added to the first line
- then the buttons were added at the bottom of the buffer
(there are more things added after the buttons,
but they haven't been changed by this commit).

- Removed the version insertion call in:
core/core-spacemacs-buffer.el

The changes above caused the version to appear twice
in the Spacemacs home buffer on startup.

I'm not sure if it has/had other uses.

This seems to be where it was first added to:
core-configuration-layer.el:
core: restore default mode line in home buffer
https://github.com/syl20bnr/spacemacs/commit/77161bd591f49107ab10ac56b33d83880fef8ce8

It was slightly modified in this commit:
core: defer distro insertion in home buffer
https://github.com/syl20bnr/spacemacs/commit/cd50d9c069ef321c778c44ff3ec975d2c15dcdbf

Here it was joined with the banner:
Fix version injection in home buffer
  Don't inject version if banner is nil
https://github.com/syl20bnr/spacemacs/commit/a764eb4eb987ee5d590756974092d45a74ab3bb4

core: condensed versions into one string in left-hand side
  spacemacs-version@emacs-version (distribution)
https://github.com/syl20bnr/spacemacs/commit/627e9344536e4524d7a3fe8609f30d34308a4977

- I had not seen that the version was joined with the
banner on purpose until now.

It seemed useful to still be able to see the versions
if one just wanted to hide the banner.

Therefore I opened this PR (it's been cherry-picked):
Uncouple version from banner in spacemacs buffer #11901
https://github.com/syl20bnr/spacemacs/pull/11901

If there's a good reason to combine them again,
then the changes can be reverted.