fix(org-present): exiting presentation always removes inline images
Spacemacs configures `org-present` to inline images while presenting. That much
makes sense. But it has a hook to turn them off unconditionally when exiting
presentation mode. This causes a problem if you just want images to show in and
out of presentation mode.

STEPS TO REPRODUCE
1. Create an Org file
2. Add a link to a file
3. Set `#+STARTUP: inlineimages`
4. `M-x org-present`
5. Exit presentation mode

EXPECTED
Because of `#+STARTUP: inlineimages` images are still inlined.

ACTUAL
The inline images have been removed upon exiting the presentation mode.

FIX
Check `org-startup-with-inline-images` on exiting presentation mode and only
remove the inline images if it is false.

TESTING
To ensure that this did not introduce a regression of the converse behavior, I
also tested with `#+STARTUP: noinlineimages`, and it still turned off the images
when exiting the presentation.