[eshell] Skip prompt text on insert
When entering insert mode, don't start editing in the middle of the prompt text.

That is, given:

~/foo/bar> some command
  ^            ^       ^
  A            B       C

Cursor A will jump to position C but cursor B won't.

Edit by duianto
This also stops the cursor from jumping to the end of the last line,
when entering insert state, from all but the last line in a multi line.

Example:
~/foo/bar>(+
1{cursor in normal state on the 1}
2)

Entering insert state: i

Before:
~/foo/bar>(+
1
2){cursor in insert state after the close parenthesis}

After:
~/foo/bar>(+
{cursor in insert state before the 1}1
2)