[eshell] Protect prompt only
If the user enters text while a command is running and the command does not
consume the text, eshell will insert this text after it displays the prompt.
Unfortunately, that happens after this function was called so this text ends up
getting protected along with the prompt text.

This change protects up until the end of the last output (the end of the
prompt text), leaving buffered input unprotected.