Add sqlfmt-region command; bindings; customize group and new setting
Enhances `sqlfmt` support in SQL layer.

`sqlfmt` was only being applied against the whole buffer, and did not allow
invocation against a specific region in the buffer. `sqlfmt` customize options
were not included under the `sql` group. It was not possible to control if
`sqlfmt` reused an existing error buffer or generated new ones.

- Added `sqlfmt-region` interactive command and binding under `sql-mode` keymap
  - `sqlfmt-buffer` calls into `sqlfmt-region` with `(point-min)` and
    `(point-max`)

- Added `sqlfmt` customize group as child to `sql` customize group

- Added `sqlfmt-reuse-error-buffer` option

- Updated `sql` layer documentation to reflect changes