Add docs about ace mode in treemacs
There is a popular question about openning file in exact window.
There is an ace mode for that, but it is not so clear how to use
it for new users.

That commmit add information how to use ace mode and how tweak it
based on information from https://github.com/Alexander-Miller/treemacs/issues/266