[scala] Avoid depending on abandoned ensime backend
Ensime seems to be finally dead, as ensime-mode is not longer
available on melpa. The same applies to ob-scala the package
which delivered scala support for org babel.

I have changed the layers default to metals and took care that
ensime is not tried to be installed until it is really selected
as a package.

In addition I have also fixed some smaller issues in the layer
which caused ensime specific settings to be forced even when
metals was selected as a backend.

I have also removed the not longer existing org-babel support
for scala as it requires ob-scala which in turn is based on ensime.

See https://github.com/hvesalai/emacs-scala-mode/issues/155 for details.