bind-map.el counterfeits ein:notebook-mode
bind-map replaces on the `minor-mode-map-alist` the keymap for
ein:notebook-mode with the evilified map, thus making the original
ein:notebook-mode-map inaccessible.

This is might be intentional as one wonders whether having active both the
original keymap and the evilified keymap makes sense.

In EIN's case, yes, we really want the original keymap, e.g., C-c C-c
to execute, to work.  Otherwise C-c C-c defaults to running a python
shell via the python-mode-map, which is really not what we want.