Fix bug with keyboard-layout layer an bépo layout
When the bépo layout is set from the `keyboard-layout` layer, the
message `(wrong-type-argument keymapp nil)` shows up. After some
digging, it turns out Spacemacs tries to set shortcuts for `nil`
`map` variables in this function.

This commit ensures it is impossible to try to set a shortcut for an
inexisting keymap.

This commit has been tested with Emacs 27.0.90.
It also removes an unused `if` statement