Wakatime: readme, note api key in bug reports
In the Wakatime layer add a line to the readme to hightlight the fact that having the api key in the main layer config will submit it as part of a bug report. There is probably a better way to achieve this but at least this highlights it for now.