[go] Make golangci run all available checkers and backup with go-build
Flycheck-golangci was not properly conigured which resulted in
not replacing disabled standard linters like go-test.

In addition there seems to be an issue in the current golangci
build causing the linter to fail. In this case spacemacs now
shows at least basic errors from go-build. When they are fixed
the linter should be able to work normally again.