Warn user they are reading develop docs
Draft commit, definitely still needs work.

The intent is to let users know they are reading documentation that might not
apply to their version of spacemacs. The default install is based on master,
but the docs for both master and develop show up in search engine results.

Currently the only way to see that you are reading the develop docs is by
looking at the subdomain. You'd also need to know that a non-develop website
exists, which is not obvious.