[javascript] add npm-mode (#14347)
* [javascript] add npm-mode

Add npm mode to the javascript layer.  This allows easy execution of npm
commands at the project root from any javascript file in the project.

* [javascript] defer loading of npm-mode

Defer loading of the npm-mode package.

* [javascript] remove superfluous `progn'

The `:init' command is only a single command and thus doesn't need to be wrapped
in a `progn'.

* [javascript] fix typo in npm-mode docs

s/avilable/available/

* [javascript] add additional npm-mode commands

Add and document missing commands from npm-mode.