FAQ on including _ in words for `*` searches
This expands the entry on including underscores and dashes in word motions to discuss `*` and `#` searches. There is a new solution for them is different, and simpler, than the solution for the other motions (but it is difficult to find).

For me, personally, this obviates the necessity of messing with the syntax definitions.

Some related issues: https://github.com/syl20bnr/spacemacs/issues/9740, 
https://github.com/syl20bnr/spacemacs/pull/1248, 
https://github.com/syl20bnr/spacemacs/issues/1261.

Many thanks to lyro@ for fixing https://github.com/emacs-evil/evil/issues/360.