[core] fix warnings about overwriting by aliasing
Fixes this warning which annoying pops up a warning buffer on the launching
IELM:

#+begin_quote
Warning (defvaralias): Overwriting value of ‘inferior-emacs-lisp-mode-hook’ by aliasing to ‘ielm-mode-hook’
#+end_quote

On the one hand, this change scares me, because it impacts a lot of modes and I
have no idea how to regression test it. Someone with deeper Spacemacs knowledge
should be able to reassure me or tell me why this is bad.

On the other hand, I cannot see how it makes sense to use a mode's hook before
the mode has even been loaded; though the problem here only seems to emerge due
to the use of an alias, in my case for ~ielm-mode~, which adds
~inferior-emacs-lisp-mode-hook~ as an alias for ~ielm-mode-hook~.