Fix #14492 don't change the selected window in delete-window
Before this commit, when delete-window was called with a window
argument, this advice changed the selected window and did not restore
it.  This could cause strange effects when the selected window was a
minibuffer window, because `one-window-p` is called with NOMINI.

Instead, we limit the dynamic extent of the selected window change to
the call to `one-window-p`.  Using `with-selected-window` also ensures
that the current buffer and selected window remain consistent after
the advice.