[auto-completion] Remove standard transformers
These try to modify the completions on a global basis.
This is causing a lot of issues with more intelligent
backends, which tend to do this in a more efficient
way already.

I have removed the global bindings for these now.
Layers should make `company-transformers` buffer local
and add specific versions of these when required.