[pdf] Change pdf-tools source to maintained fork
The original repo for pdf-tools at https://github.com/politza/pdf-tools
is no longer maintained, and its maintainer seems to be unreachable,
see the issue https://github.com/politza/pdf-tools/issues/659.

Github user vedang forked the project in order to become its new
maintainer at https://github.com/vedang/pdf-tools.

This commit changes the source of the pdf-tools package and makes it
point to the new repository. This is meant to be a workaround until
either the original maintainer of pdf-tools comes back or the repo used
by the MELPA is updated.

The recipe is based on MELPA’s recipe for pdf-tools:
https://github.com/melpa/melpa/blob/master/recipes/pdf-tools