Fix clone workspace
problem
Trying to clone a workspace `SPC l w c`
Does create a new workspace,
but it just shows the scratch buffer.

notes
In addition to fixing the cloning.

Show a minibuffer message:
Workspace: 1, cloned to Workspace: 2

To make it clear that the cloning worked.
Because the only visible difference, is that
one of the left side mode line numbers changed.