More refactor related to alist
Many times we first retrieve a key-value pair from an alist, and only when the
pair exists, we need to proceed. So it's advisable to use `when-let*` form to
instead of `let*` to bind the pair.

Another common pattern is to let-bind a key-value pair from an alist, and then
modify the value if it exist, or insert a new key-value pair if it doesn't. This
can be replaced by `(setf (alist-get KEY ALIST) VALUE)`.