[exwm] Remove manual XF86 keybindings
Recently, #14714 added `desktop-environment-mode`.  This means that we don't
need to separately define bindings like `<XF86MonBrightnessUp>`, but can instead
lean on the user to enable `desktop-environment-mode` in `init.el`.

`desktop-environment-mode` should not be loaded by default since it is
linux/unix-centric.

Also, removes a redundant binding for popping shell: instead of both `s-'` and
`s-RET`, the only command bound is `s-'`.  The README reflects this change.
The rationale is that `s-'` in not necessarily easy on non-qwerty layouts.