Ensure hjkl bindings are respected during company search/filter
At present, if using vim style (or hybrid with
`hybrid-style-enable-hjkl-bindings` set to t), then the `auto-completion` layer
will map `C-j` and `C-k` to `company-select-next` and `company-select-previous`,
respectively on the `company-active-map` keymap, and similarly binds `C-l` to
`company-complete-selection`. However, if one filters the results (for example
using `C-M-/` or by searching with `C-s`), then `C-j`, `C-k`, and `C-l` all
cause the filtering to stop and returns the completion list to what it was
pre-filtering. This seems undesirable as presumably one would want to use `C-j`,
`C-k`, and `C-l` to navigate the filtered list rather than having to switch to
arrow keys or `C-n` and `C-p`.

When searching (or having filtered), a new keymap called `company-search-map` is
active, which does not have `C-j` or `C-k` bound on it explicitly, and has a
default key mapping that causes this fallback behavior.

Add `C-j`, `C-k`, and `C-l` to the `company-search-map`. For some reason,
this same function adds `C-n` and `C-p` to the `company-active-map` when not
using vim or hybrid bindings with `hybrid-style-enable-hjkl-bindings` set to
`nil` (this seems redundant since they're already on the keymap but perhaps I
misunderstand something). This commit also does the same thing for
`company-search-map` in case it might suffer from the same issues.