[markdown] Remove obsolete orgtbl-mode cleanup
for it has a lot of side effects by globally replacing
-+- in the entire markdown buffer. Up to now markdown
has its own table insert function available via
`SPC i T` which is to be prefered.

In addition there seems to be no active
binding for `orgtbl-mode` in markdown making it very
unlikely that someone is still using it within markdown-mode.