Don't fail hard when trying to delete a non-existing package
During a package update, a package might fail installation, for example due to a
higher required emacs version. When rolling back the package update, spacemacs
tries to uninstall the new version of the package before restoring the old
version from the rollback cache. If the new version was never successfully
installed, the uninstall failed, and with it the whole rollback.

Now the uninstall error is ignored, and the rollback continues and succeeds.