[transmission] Add new keybindings (#14819)
This commit adds 3 sets of keybindings and organises the keybindings so that it
can be read easier. The README file is also organised and cleaned up to be read
easier.

1. Marking and labelling

   Add a new keybinding to mark torrents and another to set the labels at point
   or for marked torrents.

2. Queueing

   Keybindings to manage how the queueing of torrents are added using the
   uppercase vim movement keys. A keybinding to logically sort by columns is
   also added.

3. Turtle mode

   A new keybinding while in `transmission-mode` toggles the
   `transmission-turtle-mode` minor mode. When it is active, a new set of 3
   keybindings will be active to change the settings of the minor mode.

Although there is a keybinding for `transmission-remove` in `transmission-mode`,
a new keybinding for `transmission-delete` is also added. This is done since
`transmission-remove` merely removes the torrent from Transmission but does not
delete the file(s).

The rest of the changes are reflected in `CHANGELOG.develop`.